backend/requirements.txt
# FastAPI核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库相关
sqlalchemy==2.0.23
pymysql==1.1.0
aiomysql==0.2.0
aiosqlite==0.20.0
greenlet==3.0.3
alembic==1.13.1
redis==5.0.1

# AI和机器学习
# openai版本与langchain-openai需要兼容，提升至>=1.6.1
openai>=1.6.1,<2.0.0
langchain==0.0.340
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers==2.2.2
# 兼容 Python 3.12 的 NumPy 版本，避免 1.24.x 构建失败
numpy==1.26.4
pandas==2.1.4

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 工具库
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
asyncio-mqtt==0.16.1

# Chromadb 依赖补充，避免运行时导入失败
requests>=2.31.0
PyYAML>=6.0.0
typer>=0.9.0
tokenizers>=0.13.2
pypika>=0.48.9
mmh3>=4.0.1
grpcio>=1.58.0
opentelemetry-api>=1.2.0
opentelemetry-exporter-otlp-proto-grpc>=1.2.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-sdk>=1.2.0
posthog>=2.4.0
pulsar-client>=3.1.0
chroma-hnswlib==0.7.3
# chromadb 需要 bcrypt>=4.0.1
bcrypt>=4.0.1

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0